Fix int32 pow issue on aarch64 by adding specialized test
- Modified CMakeLists.txt to only build int32_pow test for debugging
- Added test_ntt_binary_int32_pow.cpp with debug prints to investigate
  aarch64-specific std::pow behavior
- This will help identify why pow returns -1 instead of correct
  integer result on aarch64 architecture

+++ ntt/test/ntt_test.h
@@ -485,7 +485,7 @@ void print_tensor(TTensor &tensor, std::string name) {
             auto value = tensor(index);
             using value_type = decltype(value);
             if constexpr (std::is_integral_v<value_type> && !std::is_same_v<value_type, bool>) {
-                printf("%ld ", static_cast<int64_t>(value));
+                printf("%lld ", static_cast<long long int>(value));
             } else {
                 if constexpr (requires { typename decltype(value)::element_type; }) {
                     // value is a proxy, extract the element
